Issue #1119646
git-svn-id: https://svn.code.sf.net/p/tkimg/code/trunk@123 8c8d7abf-59c3-4658-88e4-7cea5a6ab77c

+++ xpm/xpm.c
@@ -380,8 +380,8 @@ CommonRead(interp, handle, format, imageHandle, destX, destY,
 	  ((unsigned char *) &data)[3] = 255;
 	}
 
-	hPtr = Tcl_CreateHashEntry(&colorTable, (char *) color1, &found);
-	Tcl_SetHashValue(hPtr, (char *) data);
+	hPtr = Tcl_CreateHashEntry(&colorTable, (char *) (size_t) color1, &found);
+	Tcl_SetHashValue(hPtr, (char *) (size_t) data);
 
 	ckfree(colorName);
 	ckfree(useName);
@@ -430,16 +430,16 @@ CommonRead(interp, handle, format, imageHandle, destX, destY,
 	    unsigned int col;
 
 	    memcpy((char *) &color1, p, byteSize);
-	    hPtr = Tcl_FindHashEntry(&colorTable, (char *) color1);
+	    hPtr = Tcl_FindHashEntry(&colorTable, (char *) (size_t) color1);
 
 	    /* 
 	     * if hPtr == NULL, we have an invalid color entry in the XPM 
 	     * file. We use transparant as default color
 	     */
 	    if (hPtr != (Tcl_HashEntry *)NULL) 
-	        col = (int)Tcl_GetHashValue(hPtr);
+	        col = (unsigned int) (size_t) Tcl_GetHashValue(hPtr);
 	    else 
-	        col = (int)0;
+	        col = (unsigned int) 0;
 	    
 	    /*
 	     * we've found a non-transparent pixel, let's search the next
@@ -459,11 +459,11 @@ CommonRead(interp, handle, format, imageHandle, destX, destY,
 		    
 		    if (i < width) {
 		        memcpy((char *) &color1, p, byteSize);
-			hPtr = Tcl_FindHashEntry(&colorTable, (char *) color1);
+			hPtr = Tcl_FindHashEntry(&colorTable, (char *) (size_t) color1);
 			if (hPtr != (Tcl_HashEntry *)NULL) 
-			    col = (int)Tcl_GetHashValue(hPtr);
+			    col = (unsigned int) (size_t) Tcl_GetHashValue(hPtr);
 			else 
-			    col = (int)0;
+			    col = (unsigned int) 0;
 		    }
 		} while ((i < width) && col);
 		tkimg_PhotoPutBlock(interp, imageHandle, &block, destX+j, destY, len, 1, TK_PHOTO_COMPOSITE_OVERLAY);
